Strip out unnecessary changes as a todo
