+++ src/index.ts
@@ -3,11 +3,11 @@ export { default as useCioPlp } from './hooks/useCioPlp';
 export { default as useSearchResults } from './hooks/useSearchResults';
 export { default as useBrowseResults } from './hooks/useBrowseResults';
 export { default as useCioClient } from './hooks/useCioClient';
-export { useCioPlpContext } from './hooks/useCioPlpContext';
 export * from './hooks/callbacks';
 
 // Components
 export { default as CioPlp } from './components/CioPlp';
 export { default as ProductCard, ProductCardProps } from './components/ProductCard';
+export { CioPlpContext } from './PlpContext';
 
 export * from './types';
