+++ src/stories/components/ProductCard/ProductCardExample.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import ProductCard, { ProductCardProps } from '../../../components/ProductCard';
 import { DEMO_API_KEY } from '../../../constants';
-import CioPlp from '../../../components/CioPlp';
+import { CioPlpContext } from '../../../PlpContext';
 
 /**
  * A Product Card UI Component
  */
 export default function ProductCardExample({ ...props }: ProductCardProps) {
   return (
-    <CioPlp apiKey={DEMO_API_KEY}>
+    <CioPlpContext apiKey={DEMO_API_KEY}>
       <ProductCard {...props} />
-    </CioPlp>
+    </CioPlpContext>
   );
 }
