+++ src/stories/hooks/useSearchResults/Documentation.mdx
@@ -7,7 +7,7 @@ import StateDocs from './StateDocs.md?raw';
 ## `useSearchResults`
 A React Hook to retrieve search results using Constructor.
 
->Note: useSearchResults can be only called in components that are direct children of `<CioPlp />`.
+>Note: useSearchResults can be only called in components that are direct children of `<CioPlpContext />`.
 
 
 {/* -------- STATE -------- */}
