+++ src/stories/hooks/useSearchResults/Usage Examples.mdx
@@ -19,9 +19,9 @@ function myApp() {
   ...
   return (
     <>
-    <CioPlp apiKey={MY_API_KEY}>
+    <CioPlpContext apiKey={MY_API_KEY}>
       <SearchResults query={query} />
-    </CioPlp>
+    </CioPlpContext>
     </>
   )
 }
